refactor: create separate controllers for rendering views & move api controllers to separate directory
