+++ app/Http/Controllers/CityViewController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class CityViewController extends Controller
+{
+    public function index()
+    {
+        return view('cities');
+    }
+}
